扩大BlockingCollection的容量

+++ Frameworks/Server/Server.ProcessorManager.cs
@@ -62,7 +62,7 @@ namespace GoPlay.Services
             foreach (var processor in Processors)
             {   
                 var name = processor.GetName();
-                m_packageQueues[name] = new BlockingCollection<Package>(byte.MaxValue);
+                m_packageQueues[name] = new BlockingCollection<Package>(ushort.MaxValue);
                 m_broadcastQueues[name] = new BlockingCollection<(uint, int, object)>(byte.MaxValue);
                 m_tasks[name] = TaskUtil.LongRun(() => PackageLoop(processor, m_cancelSource.Token), m_cancelSource.Token);
             }
